rename bitmask
